Purchasing an item checks if in shop and if the client has the gold. Added item to inventory hud on purchase

+++ Source/MobaProject/Public/Items/BaseItem.h
@@ -47,7 +47,7 @@ public:
 	FText Description;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Item")
-	int32 ItemCost = 0;
+	float ItemCost = 0;
 
 	//The tooltip in the inventory for this item
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category="Item")
